small refactoring customer account managing cards, changed names of routes
