+++ view/frontend/layout/omise_manage_cards.xml
@@ -1,9 +1,9 @@
-<?xml version="1.0"?>s
+<?xml version="1.0"?>
+
 <page
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../../../../../lib/internal/Magento/Framework/View/Layout/etc/page_configuration.xsd">
     <update handle="customer_account"/>
     <body>
-
         <referenceContainer name="content">
             <block class="Omise\Payment\Block\Customer\Cards" name="omise.cards" template="Omise_Payment::cards.phtml"
             cacheable="false"></block>
